backend/requirements.txt
Flask==1.1.2
Flask-Caching==1.8.0
xmltodict==0.12.0
redis==3.5.2
requests==2.23.0
markdown
python-dateutil==2.8.1
ruamel.yaml==0.16.10
transliterate==1.10.2
gitpython==3.1.3
